CheckMechBackend/apps/core/serializers.py
from rest_framework import serializers
from .models import CarBrand, ProfilePhoto


class CarBrandSerializer(serializers.ModelSerializer):
    class Meta:
        model = CarBrand
        fields = "__all__"


class ProfilePhotoSerializer(serializers.ModelSerializer):
    class Meta:
        model = ProfilePhoto
        fields = "__all__"