CheckMechBackend/requirements.txt
django
djangorestframework
djoser 
djangorestframework-simplejwt
drf_yasg
gunicorn
django-cors-headers
drf-nested-routers
pillow